% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eq.r
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Tidy and subset earthquake data}
\usage{
eq_clean_data(data)
}
\arguments{
\item{data}{Name of input data frame on which to apply function.}
}
\value{
This function returns a data frame table containing a cleansed subset of the earthquake data.
}
\description{
This function (eq_clean_data) removes unnecessary columns using the select_ function from the dplyr package (\code{\link[dplyr]{select_}}).
The columns that are kept in the data are: YEAR, MONTH, DAY, LATITUDE, LONGITUDE, LOCATION_NAME, TOTAL_DEATHS, COUNTRY, EQ_PRIMARY.
Earthquakes with negative year values are excluded from the data using the filter_ function from the dplyr package (\code{\link[dplyr]{filter_}}).
The mutate_ function from the dplyr package (\code{\link[dplyr]{mutate_}}) is used to create a date column and ensure analysis columns are numeric.
}
\examples{
\dontrun{
eq_clean <- eq_clean_data(eq_data)
eq_clean <- eq_clean_data(data= eq_data)
}

}
