% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eq.r
\name{eq_create_label}
\alias{eq_create_label}
\title{Create an interactive map of earthquake data with labels}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{Name of input data frame on which to apply function.}
}
\value{
This function creates labels with formatting for the earthquake data.
}
\description{
This function (eq_create_label) creates and applies formatting to the labels for the points on the interactive earthquake map.
}
\examples{
\dontrun{

eq_map_tl <- eq_read_data("signif.txt") \%>\% 
eq_clean_data() \%>\%
eq_location_clean() \%>\%
dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\% 
dplyr::mutate(popup_text = eq_create_label(.)) \%>\% 
eq_map(annot_col = "popup_text")

readr::read_delim("signif.txt", delim = "\\t") \%>\% 
eq_clean_data() \%>\% 
dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\% 
dplyr::mutate(popup_text = eq_create_label(.)) \%>\% 
eq_map(annot_col = "popup_text")
}

}
