% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eq.r
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Tidy location names field in earthquake data}
\usage{
eq_location_clean(data)
}
\arguments{
\item{data}{Name of input data frame on which to apply function.}
}
\value{
This function returns a data frame table where the LOCATION_NAME field in the earthquake data is cleansed.
}
\description{
This function (eq_location_clean) removes country names at the beginning of the LOCATION_NAME field.
It is assumed that country names are followed by a colon when a subsequent city name is provided.
Where only the country name is given in the LOCATION_NAME field, the country name is removed since this information is in the country field.
The removal of country names is performed using the mutate_ function from the dplyr package (\code{\link[dplyr]{mutate_}}).
The LOCATION_NAME field is converted to title case using the str_to_title function from the stringr package (\code{\link[stringr]{str_to_title}}).
}
\examples{
\dontrun{
eq_clean_loc <- eq_location_clean(eq_clean)
eq_clean_loc <- eq_location_clean(data= eq_clean)
}

}
