% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eq.r
\name{eq_map}
\alias{eq_map}
\title{Create an interactive map of earthquake data with labels}
\usage{
eq_map(data, annot_col)
}
\arguments{
\item{data}{Name of input data frame on which to apply function.}

\item{annot_col}{Name of field containing label information.}
}
\value{
This function returns an interactive map of the earthquake data.
}
\description{
This function (eq_map) creates an interactive map of earthquake data using the leaflet function from the leaflet package (\code{\link[leaflet]{leaflet}}).
}
\examples{
\dontrun{

eq_map_tl <- eq_read_data("signif.txt") \%>\% 
eq_clean_data() \%>\%
dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\% 
eq_map(annot_col = "DATE")

readr::read_delim("signif.txt", delim = "\\t") \%>\% 
eq_clean_data() \%>\% 
dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\% 
eq_map(annot_col = "DATE")
}

}
