% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eq.r
\docType{data}
\name{geomtimeline}
\alias{geomtimeline}
\title{Geom for plotting timeline of earthquake data}
\format{An object of class \code{geomtimeline} (inherits from \code{Geom}, \code{ggproto}) of length 6.}
\usage{
geomtimeline
}
\description{
This geom (geomtimeline) subsets the data to values between xmin and xmax using the filter_ function from the dplyr package (\code{\link[dplyr]{filter_}}).
The line elements of the graph are created using the polylineGrob function from the grid package (\code{\link[grid]{polylineGrob}}).
The point elements of the graph are created using the pointsGrob function from the grid package (\code{\link[grid]{pointsGrob}}).
The line and point elements are plotted together using the gList function from the grid package (\code{\link[grid]{gList}}).
Returns a timeline showing eathquake sizes and intensities between specified dates.
Required_aes A character vector of aesthetics needed to render the geom; Date field - x, lowest date to include - xmin, highest date to include - xmax.
Default_aes A list generated by aes() of default values for aesthetics; Group variable where separate timelines should be created for each unique value - y. 
Variable where values distinguished by point outline colour - colour. 
Variable where values distinguished by point fill colour - fill, variable where values distinguished by point size - size.
Transparency of points - alpha, shape of points - shape, width of border to points - stroke.
Setup_data Called once for each layer, after setp_params(); Returns modified data between specified dates.
Draw_key Renders a single legend key.
Draw_group Function to transform the data to appropriate graphing coordinate system.
Not a user facing element of package.
}
\keyword{datasets}
