% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eq.r
\docType{data}
\name{geomtimelinelabel}
\alias{geomtimelinelabel}
\title{Geom for plotting timeline labels for earthquake data timelines}
\format{An object of class \code{geomtimelinelabel} (inherits from \code{Geom}, \code{ggproto}) of length 6.}
\usage{
geomtimelinelabel
}
\description{
This geom (geomtimelinelabel) subsets the data to values between xmin and xmax using the filter_ function from the dplyr package (\code{\link[dplyr]{filter_}}).
If applicable the data is grouped by the y variable using the group_by_ function from the dplyr package (\code{\link[dplyr]{group_by_}}).  
The top n_max values of the size variable are kept for each group (if applicable) using the top_n function from the dplyr package (\code{\link[dplyr]{top_n}}).
The data is ungrouped (if applicable) using the ungroup function from the dplyr package (\code{\link[dplyr]{ungroup}}).
The line elements of the graph are created using the polylineGrob function from the grid package (\code{\link[grid]{polylineGrob}}).
The text elements of the graph are created using the textGrob function from the grid package (\code{\link[grid]{textGrob}}).
The line and text elements are plotted together using the gList function from the grid package (\code{\link[grid]{gList}}).
Returns a timeline label showing information regarding eathquake sizes, intensities, and locations between specified dates.
Required_aes A character vector of aesthetics needed to render the geom; Date field - x, lowest date to include - xmin. 
Highest date to include - xmax, variable holding labels - label.
Default_aes A list generated by aes() of default values for aesthetics; Group variable where separate timelines should be created for each unique value - y.  
Variable where values distinguished by point size - size.
Setup_data Called once for each layer, after setp_params(); Returns modified data between specified dates.
Draw_key Renders a single legend key.
Draw_panel Function to transform the data to appropriate graphing coordinate system.
Not a user facing element of package.
}
\keyword{datasets}
